# clients/go/api/openapi.yaml
openapi: 3.0.2
info:
  title: FastAPI
  version: 0.1.0
servers:
- url: /
paths:
  /:
    get:
      operationId: ready__get
      responses:
        "200":
          content:
            application/json:
              schema:
                title: Response Ready  Get
                type: string
          description: Successful Response
      summary: Ready
  /translate:
    get:
      operationId: translate_translate_get
      parameters:
      - description: The text to be translated
        explode: true
        in: query
        name: source_text
        required: true
        schema:
          description: The text to be translated
          title: Source Text
          type: string
        style: form
      - description: The ISO-639-1 code of the language to translate the text to
        explode: true
        in: query
        name: to_language
        required: true
        schema:
          description: The ISO-639-1 code of the language to translate the text to
          maxLength: 2
          title: To Language
          type: string
        style: form
      - description: The ISO-639-1 code of the language to translate the text from
          - if notspecified then detection will be attempted
        explode: true
        in: query
        name: from_language
        required: false
        schema:
          description: The ISO-639-1 code of the language to translate the text from
            - if notspecified then detection will be attempted
          maxLength: 2
          title: From Language
          type: string
        style: form
      - description: Which translation engine to use. Choices are microsoft, google,
          amazon, papago, deepl, yandex and best
        explode: true
        in: query
        name: preferred_engine
        required: false
        schema:
          default: best
          description: Which translation engine to use. Choices are microsoft, google,
            amazon, papago, deepl, yandex and best
          title: Preferred Engine
          type: string
        style: form
      - description: Whether to return word alignment information or not
        explode: true
        in: query
        name: with_alignment
        required: false
        schema:
          default: false
          description: Whether to return word alignment information or not
          title: With Alignment
          type: boolean
        style: form
      - description: Whether to fallback to the best available engine if the preferred
          engine does not succeed
        explode: true
        in: query
        name: fallback
        required: false
        schema:
          default: false
          description: Whether to fallback to the best available engine if the preferred
            engine does not succeed
          title: Fallback
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Translate
    post:
      operationId: translate_post_translate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Translate Post
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    TranslationRequest:
      example:
        with_alignment: false
        source_text: source_text
        from_language: from_language
        preferred_engine: best
        to_language: to_language
        fallback: false
      properties:
        source_text:
          description: The text to be translated
          title: Source Text
          type: string
        to_language:
          description: The ISO-639-1 code of the language to translate the text to
          maxLength: 2
          title: To Language
          type: string
        from_language:
          description: The ISO-639-1 code of the language to translate the text from
            - if notspecified then detection will be attempted
          maxLength: 2
          title: From Language
          type: string
        preferred_engine:
          default: best
          description: Which translation engine to use. Choices are microsoft, google,
            amazon, papago, deepl, yandex and best
          title: Preferred Engine
          type: string
        with_alignment:
          default: false
          description: Whether to return word alignment information or not
          title: With Alignment
          type: boolean
        fallback:
          default: false
          description: Whether to fallback to the best available engine if the preferred
            engine does not succeed
          title: Fallback
          type: boolean
      required:
      - source_text
      - to_language
      title: TranslationRequest
      type: object
    TranslationResponse:
      example:
        translated_text: 안녕하세요
        engine: microsoft
        engine_version: "3.0"
        from_language: en
        to_language: ko
        source_text: hello
        alignment:
        - dest:
            end: "4"
            start: "0"
            text: 안녕하세요
          src:
            end: "4"
            start: "0"
            text: hello
      properties:
        engine:
          title: Engine
          type: string
        engine_version:
          title: Engine Version
          type: string
        detected_language_confidence:
          title: Detected Language Confidence
          type: number
        from_language:
          title: From Language
          type: string
        to_language:
          title: To Language
          type: string
        source_text:
          title: Source Text
          type: string
        translated_text:
          title: Translated Text
          type: string
        alignment:
          items:
            additionalProperties:
              additionalProperties:
                type: string
              type: object
            type: object
          title: Alignment
          type: array
      required:
      - engine
      - engine_version
      - from_language
      - source_text
      - to_language
      - translated_text
      title: TranslationResponse
      type: object
    ValidationError:
      properties:
        loc:
          items:
            type: string
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object